ci: use common script for tests on travis and gitlab
For one, it's not unreasonable that we want to run the same
tests both for gitlab and travis.

Move the actual tests into a script, which is spawned by both.

We still can do something different, based on the environment.
The advantage here is, that the common part will be shared, and
the places where we differ will be easier to recognize.